+++ lib/article_json/configuration.rb
@@ -52,8 +52,8 @@ module ArticleJSON
     # @param [Symbol] exporter
     # @param [Hash[Symbol => Class]] type_class_mapping
     def register_element_exporters_for(exporter, type_class_mapping)
-      unless %i(html).include?(exporter)
-        raise ArgumentError, '`exporter` only supports `:html` '\
+      unless %i(html amp).include?(exporter)
+        raise ArgumentError, '`exporter` needs to be either `:html` or `:amp` '\
                              "but is `#{exporter.inspect}`"
       end
       if !type_class_mapping.is_a?(Hash) ||