+++ lib/article_json/utils/o_embed_resolver/tweet.rb
@@ -2,18 +2,22 @@ module ArticleJSON
   module Utils
     module OEmbedResolver
       class Tweet < Base
+        # Human readable name of the resolver
+        # @return [String]
+        def name
+          'Tweet'
+        end
+
         # The URL for the oembed API call
         # @return [String]
         def oembed_url
           'https://api.twitter.com/1/statuses/oembed.json?align=center' \
-            "&url=#{tweet_url}"
+            "&url=#{source_url}"
         end
-
-        private
 
         # The URL of the tweet
         # @return [String]
-        def tweet_url
+        def source_url
           handle, tweet_id = @element.embed_id.split('/', 2)
           "https://twitter.com/#{handle}/status/#{tweet_id}"
         end