+++ lib/article_json/utils/o_embed_resolver/vimeo_video.rb
@@ -2,17 +2,21 @@ module ArticleJSON
   module Utils
     module OEmbedResolver
       class VimeoVideo < Base
+        # Human readable name of the resolver
+        # @return [String]
+        def name
+          'Vimeo video'
+        end
+
         # The URL for the oembed API call
         # @return [String]
         def oembed_url
-          "https://vimeo.com/api/oembed.json?url=#{video_url}"
+          "https://vimeo.com/api/oembed.json?url=#{source_url}"
         end
-
-        private
 
         # The video URL of the element
         # @return [String]
-        def video_url
+        def source_url
           "https://vimeo.com/#{@element.embed_id}"
         end
       end