+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ? - WIP
 - Configurable HTML element exporters
 - Resolve oembed elements in HTML export
+- Export quotes as `<div>` instead of `<aside>`
 
 ## 0.1.0 - 2017/09/20
 This is the very first release, with the following functionality: