+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Resolve oembed elements in HTML export
 - Export quotes as `<div>` instead of `<aside>`
 - Make Google Parser more fault tolerant
+- Support Vimeo videos with old flash player URLs
 
 ## 0.1.0 - 2017/09/20
 This is the very first release, with the following functionality: